test_imports.py
import transformers
import datasets
import pandas as pd
import numpy as np
import sklearn

import evaluate
import nltk
import scipy
from bert_score import score

import openai
import google.generativeai as genai
from huggingface_hub import HfApi

from bnlp import NLTKTokenizer

print("✅ All imports successful!")