requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
appdirs==1.4.4
argcomplete==3.6.3
asgiref==3.11.1
attr==0.3.1
attrs==26.1.0
azure-core==1.39.0
azure-storage-blob==12.28.0
bert-score==0.3.13
black==26.3.1
bleach==5.0.1
bnlp_toolkit==4.4.1
boto3==1.42.84
botocore==1.42.84
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.2
colorama==0.4.6
contourpy==1.3.3
croniter==6.2.2
cryptography==46.0.6
cuda-bindings==13.2.0
cuda-pathfinder==1.5.2
cuda-toolkit==13.0.2
cycler==0.12.1
datamodel-code-generator==0.26.1
datasets==4.8.4
defusedxml==0.7.1
dill==0.4.1
distro==1.9.0
Django==5.1.15
django-annoying==0.10.6
django-cors-headers==4.7.0
django-csp==3.7
django-debug-toolbar==3.2.1
django-environ==0.10.0
django-extensions==3.2.3
django-filter==24.3
django-migration-linter==5.2.0
django-ranged-fileresponse==0.1.2
django-rq==3.1
django-storages==1.12.3
django-user-agents==0.4.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.1
dnspython==2.8.0
drf-dynamic-fields==0.3.0
drf-flex-fields==0.9.5
drf-generators==0.3.0
drf-spectacular==0.28.0
editdistpy==0.2.0
email-validator==2.3.0
emoji==2.15.0
evaluate==0.4.6
expiringdict==1.2.2
Faker==40.13.0
filelock==3.25.2
fonttools==4.62.1
frozenlist==1.8.0
fsspec==2026.2.0
ftfy==6.3.1
gensim==4.4.0
genson==1.3.0
google-ai-generativelanguage==0.6.15
google-api-core==2.30.2
google-api-python-client==2.193.0
google-auth==2.49.1
google-auth-httplib2==0.3.1
google-cloud-appengine-logging==1.9.0
google-cloud-audit-log==0.5.0
google-cloud-core==2.5.1
google-cloud-logging==3.15.0
google-cloud-storage==3.10.1
google-crc32c==1.8.0
google-generativeai==0.8.6
google-resumable-media==2.8.2
googleapis-common-protos==1.74.0
grpc-google-iam-v1==0.14.4
grpcio==1.80.0
grpcio-status==1.71.2
h11==0.16.0
hf-xet==1.4.3
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
huggingface_hub==1.9.1
idna==3.11
ijson==3.5.0
importlib_metadata==8.7.1
inflect==5.6.2
inflection==0.5.1
isodate==0.7.2
isort==5.13.2
Jinja2==3.1.6
jiter==0.13.0
jmespath==1.1.0
joblib==1.5.3
jsf==0.11.2
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kiwisolver==1.5.0
label-studio==1.23.0
label-studio-sdk==2.0.18
launchdarkly-server-sdk==8.2.1
lxml==6.0.2
lxml_html_clean==0.4.4
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
mdurl==0.1.2
mpmath==1.3.0
multidict==6.7.1
multiprocess==0.70.19
mypy_extensions==1.1.0
networkx==3.6.1
nltk==3.9.4
numpy==2.4.4
nvidia-cublas==13.1.0.3
nvidia-cuda-cupti==13.0.85
nvidia-cuda-nvrtc==13.0.88
nvidia-cuda-runtime==13.0.96
nvidia-cudnn-cu13==9.19.0.56
nvidia-cufft==12.0.0.61
nvidia-cufile==1.15.1.6
nvidia-curand==10.4.0.35
nvidia-cusolver==12.0.4.66
nvidia-cusparse==12.6.3.3
nvidia-cusparselt-cu13==0.8.0
nvidia-nccl-cu13==2.28.9
nvidia-nvjitlink==13.0.88
nvidia-nvshmem-cu13==3.4.5
nvidia-nvtx==13.0.85
openai==1.109.1
opencv-python-headless==4.13.0.92
opentelemetry-api==1.40.0
ordered-set==4.0.2
packaging==26.0
pandas==3.0.2
pathspec==1.0.4
pillow==12.2.0
platformdirs==4.9.4
propcache==0.4.1
proto-plus==1.27.2
protobuf==5.29.6
psycopg==3.3.3
psycopg-binary==3.3.3
pyarrow==22.0.0
pyasn1==0.6.3
pyasn1_modules==0.4.2
pyboxen==1.3.0
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.20.0
PyJWT==2.12.1
pyparsing==3.3.2
pyRFC3339==2.1.0
python-crfsuite==0.9.12
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-json-logger==2.0.4
pytokens==0.4.1
pytz==2022.7.1
PyYAML==6.0.3
redis==5.2.1
referencing==0.37.0
regex==2026.4.4
requests==2.32.5
requests-file==3.0.1
requests-mock==1.12.1
rich==14.3.3
rpds-py==0.30.0
rq==2.6.1
rstr==3.2.2
rules==3.4
s3transfer==0.16.0
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.1
semver==3.0.4
sentencepiece==0.2.1
sentry-sdk==2.57.0
setuptools==81.0.0
shellingham==1.5.4
six==1.17.0
sklearn-crfsuite==0.5.0
smart_open==7.5.1
sniffio==1.3.1
sqlparse==0.5.5
sympy==1.14.0
symspellpy==6.9.0
tabulate==0.10.0
threadpoolctl==3.6.0
tldextract==5.3.1
together==2.7.0
tokenizers==0.22.2
toml==0.10.2
torch==2.11.0
tqdm==4.67.3
transformers==5.5.0
triton==3.6.0
typer==0.24.1
types-PyYAML==6.0.12.20250915
types-requests==2.33.0.20260402
types-tabulate==0.10.0.20260308
types-tqdm==4.67.3.20260402
typing-inspection==0.4.2
typing_extensions==4.15.0
ua-parser==1.0.2
ua-parser-builtins==202603
ujson==5.12.0
uritemplate==4.2.0
urllib3==2.6.3
user-agents==2.2.0
uuid_utils==0.14.1
wcwidth==0.6.0
webencodings==0.5.1
wheel==0.46.3
wrapt==2.1.2
xmljson==0.2.1
xxhash==3.6.0
yarl==1.23.0
zipp==3.23.0
